Flyttet config og configdata-klasser som er felles mellom ebms-provider og ebms-async til fellesmodulen.
