+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/configuration/Config.kt
@@ -1,5 +1,9 @@
 package no.nav.emottak.ebms.async.configuration
 
+import no.nav.emottak.config.KafkaPayloadProducer
+import no.nav.emottak.config.KafkaPayloadReceiver
+import no.nav.emottak.config.KafkaSignalProducer
+import no.nav.emottak.config.KafkaSignalReceiver
 import no.nav.emottak.utils.config.EventLogging
 import no.nav.emottak.utils.config.Kafka
 import no.nav.emottak.utils.environment.getEnvVar
@@ -21,26 +25,6 @@ data class Config(
     val kafkaPayloadReceiver: KafkaPayloadReceiver,
     val kafkaPayloadProducer: KafkaPayloadProducer,
     val kafkaErrorQueue: KafkaErrorQueue
-)
-
-data class KafkaSignalReceiver(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaSignalProducer(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaPayloadReceiver(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaPayloadProducer(
-    val active: Boolean,
-    val topic: String
 )
 
 data class KafkaErrorQueue(
