+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/configuration/Configurator.kt
@@ -8,6 +8,7 @@ fun config() = ConfigLoader.builder()
     .addEnvironmentSource()
     .addResourceSource("/application-personal.conf", optional = true)
     .addResourceSource("/kafka_common.conf")
+    .addResourceSource("/ebms_kafka_queues.conf")
     .addResourceSource("/application.conf")
     .withExplicitSealedTypes()
     .build()
