+++ ebms-async/src/main/resources/application.conf
@@ -15,26 +15,6 @@ kafka {
   groupId = "ebms-async"
 }
 
-kafkaSignalReceiver {
-  active = "${EBMS_SIGNAL_RECEIVER:-false}"
-  topic = "team-emottak.smtp.in.ebxml.signal"
-}
-
-kafkaSignalProducer {
-  active = "${EBMS_SIGNAL_PRODUCER:-false}"
-  topic = "team-emottak.smtp.out.ebxml.signal"
-}
-
-kafkaPayloadReceiver {
-  active = "${EBMS_PAYLOAD_RECEIVER:-false}"
-  topic = "team-emottak.smtp.in.ebxml.payload"
-}
-
-kafkaPayloadProducer {
-  active = "${EBMS_PAYLOAD_PRODUCER:-false}"
-  topic = "team-emottak.smtp.out.ebxml.payload"
-}
-
 kafkaErrorQueue {
   active = "${EBMS_RETRY_QUEUE:-false}"
   topic = "team-emottak.ebxml.retry"
