+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/configuration/Config.kt
@@ -1,5 +1,9 @@
 package no.nav.emottak.ebms.configuration
 
+import no.nav.emottak.config.KafkaPayloadProducer
+import no.nav.emottak.config.KafkaPayloadReceiver
+import no.nav.emottak.config.KafkaSignalProducer
+import no.nav.emottak.config.KafkaSignalReceiver
 import no.nav.emottak.utils.config.EventLogging
 import no.nav.emottak.utils.config.Kafka
 
@@ -10,24 +14,4 @@ data class Config(
     val kafkaSignalProducer: KafkaSignalProducer,
     val kafkaPayloadReceiver: KafkaPayloadReceiver,
     val kafkaPayloadProducer: KafkaPayloadProducer
-)
-
-data class KafkaSignalReceiver(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaSignalProducer(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaPayloadReceiver(
-    val active: Boolean,
-    val topic: String
-)
-
-data class KafkaPayloadProducer(
-    val active: Boolean,
-    val topic: String
 )
