+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/configuration/Configurator.kt
@@ -10,6 +10,7 @@ fun config() = ConfigLoader.builder()
     .addEnvironmentSource()
     .addResourceSource("/application-personal.conf", optional = true)
     .addResourceSource("/kafka_common.conf")
+    .addResourceSource("/ebms_kafka_queues.conf")
     .addResourceSource("/application.conf")
     .withExplicitSealedTypes()
     .build()
