+++ ebms-provider/src/main/resources/application.conf
@@ -6,23 +6,3 @@ kafka {
 eventLogging {
   eventLoggingProducerActive = "true"
 }
-
-kafkaSignalReceiver {
-  active = "${EBMS_SIGNAL_RECEIVER:-false}"
-  topic = "team-emottak.smtp.in.ebxml.signal"
-}
-
-kafkaSignalProducer {
-  active = "${EBMS_SIGNAL_PRODUCER:-false}"
-  topic = "team-emottak.smtp.out.ebxml.signal"
-}
-
-kafkaPayloadReceiver {
-  active = "${EBMS_PAYLOAD_RECEIVER:-false}"
-  topic = "team-emottak.smtp.in.ebxml.payload"
-}
-
-kafkaPayloadProducer {
-  active = "${EBMS_PAYLOAD_PRODUCER:-false}"
-  topic = "team-emottak.smtp.out.ebxml.payload"
-}
