+++ felles/src/main/resources/ebms_kafka_queues.conf
@@ -1,0 +1,20 @@
+
+kafkaSignalReceiver {
+  active = "${EBMS_SIGNAL_RECEIVER:-false}"
+  topic = "team-emottak.smtp.in.ebxml.signal"
+}
+
+kafkaSignalProducer {
+  active = "${EBMS_SIGNAL_PRODUCER:-false}"
+  topic = "team-emottak.smtp.out.ebxml.signal"
+}
+
+kafkaPayloadReceiver {
+  active = "${EBMS_PAYLOAD_RECEIVER:-false}"
+  topic = "team-emottak.smtp.in.ebxml.payload"
+}
+
+kafkaPayloadProducer {
+  active = "${EBMS_PAYLOAD_PRODUCER:-false}"
+  topic = "team-emottak.smtp.out.ebxml.payload"
+}
